requirements.txt
psycopg2-binary
python-dotenv
asyncpg
flake8
autopep8
bcrypt
senvdatareader